core/analyzer.py
import cv2
import numpy as np

class TestStripAnalyzer:
    def __init__(self, calibration_data=None):
        self.calibration_data = calibration_data or {}

    def preprocess_image(self, image_path):
        image = cv2.imread(image_path)
        return cv2.cvtColor(image, cv2.COLOR_BGR2HSV)

    def extract_color_zones(self, hsv_image):
        height, width, _ = hsv_image.shape
        zones = []
        for i in range(5):  # Assume 5 test zones
            x = int(width * (i + 0.5) / 5)
            zone = hsv_image[:, x-5:x+5]
            avg_color = np.mean(zone, axis=(0, 1))
            zones.append(avg_color)
        return zones

    def interpret_zones(self, zones):
        results = {}
        for i, color in enumerate(zones):
            results[f"Zone {i+1}"] = self._match_color_to_chemical(color)
        return results

    def _match_color_to_chemical(self, hsv_color):
        # Placeholder logic
        h, s, v = hsv_color
        if h < 30:
            return "Low pH"
        elif h < 60:
            return "Balanced"
        else:
            return "High Chlorine"
