requirements.txt
absl-py==2.2.2
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiohttp-retry==2.9.1
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
astroid==3.3.10
astunparse==1.6.3
async-timeout==5.0.1
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
certifi==2025.4.26
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
comtypes==1.4.10
contourpy==1.3.2
cryptography==44.0.3
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
Deprecated==1.2.18
dill==0.4.0
docopt==0.6.2
et_xmlfile==2.0.0
exceptiongroup==1.2.2
fastapi==0.115.12
Flask==3.1.1
flatbuffers==25.2.10
fonttools==4.57.0
fpdf==1.7.2
frozenlist==1.6.0
gast==0.6.0
google-pasta==0.2.0
grpcio==1.71.0
h11==0.16.0
h5py==3.13.0
idna==3.10
iniconfig==2.1.0
isort==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.0
Js2Py==0.74
keras==3.9.2
kiwisolver==1.4.8
libclang==18.1.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.3
mccabe==0.7.0
mdurl==0.1.2
ml_dtypes==0.5.1
multidict==6.4.3
mysql-connector-python==9.3.0
namex==0.0.9
numpy==2.1.3
opencv-python==4.11.0.86
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.15.0
packaging==25.0
paho-mqtt==2.1.0
pandas==2.2.3
pillow==11.2.1
platformdirs==4.3.8
pluggy==1.6.0
-e git+https://github.com/valcalepi/Deep-Blue-Pool-Chemistry.git@aa56633d19a8c39eb85143f40672490b59446379#egg=pool_controller
propcache==0.3.1
protobuf==5.29.4
PyAudio==0.2.14
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
pyFirmata==1.1.0
PyGithub==2.6.1
Pygments==2.19.1
pyjsparser==2.7.1
PyJWT==2.10.1
pylint==3.3.7
PyNaCl==1.5.0
pyparsing==3.2.3
pypiwin32==223
PyPrind==2.11.3
pyserial==3.5
pySmartDL==1.3.4
pytest==8.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pyttsx3==2.98
pytz==2025.2
pywin32==310
qrcode==8.2
reportlab==4.4.0
requests==2.32.3
rich==14.0.0
schedule==1.2.2
scikit-learn==1.6.1
scipy==1.15.3
seaborn==0.13.2
secure-smtplib==0.1.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SpeechRecognition==3.14.2
starlette==0.46.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
threadpoolctl==3.6.0
tomli==2.2.1
tomlkit==0.13.2
ttkthemes==3.2.2
twilio==9.6.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
uvicorn==0.34.2
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.2
XlsxWriter==3.2.3
yarl==1.20.0
